vote and order projection fixed

+++ Booklandbackend/src/main/java/com/bookland/demobookland/model/Order.java
@@ -35,7 +35,7 @@ public class Order {
     private Integer addressId;
 
     @Column(name = "TotalAmount", nullable = false)
-    private Integer totalAmount;
+    private Float totalAmount;
 
     @Column(name = "CampaignId")
     private Integer campaignId;
